% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtils.R
\name{cluster_library}
\alias{cluster_library}
\title{Load packages on a cluster}
\usage{
cluster_library(cl, packages)
}
\arguments{
\item{cl}{cluster object created by \link{[snow]{makeCluster}}}

\item{packages}{list or vector of package names}
}
\value{
TRUE if all packages could be loaded on all cluster nodes
}
\description{
Load packages on a cluster
}
