% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFactorCopula.R
\name{config_factor}
\alias{config_factor}
\title{Configurate the latent variables of a factorcopula model}
\usage{
config_factor(..., par = c())
}
\arguments{
\item{...}{One or more expresssions separated by commas. The name of the expression arguments have to be
a valid random number generators, the expressions have to be lists of named unquoted arguments.#'}

\item{par}{a character vector of parameter names used in the specification of the factor matrix}
}
\value{
A list of \link[rlang]{quosure}s to be used in \link[factorcopula]{fc_create} or \link[factorcopula]{fc_fit}
}
\description{
Configurate the latent variables of a factorcopula model
}
