% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFactorCopula.R
\name{factorCopula}
\alias{factorCopula}
\title{Simulate values from a factor copula model}
\usage{
factorCopula(beta, N, Z, eps, zFixed = FALSE, epsFixed = zFixed, S = NULL)
}
\arguments{
\item{beta}{a character matrix of size [NxK] indicating the names and position of the beta parameters (see example)}

\item{N}{the number of variables}

\item{Z}{a named list of size K and names equal to a random number generator function which takes an argument n and possibly other arguments passed as numeric values or characters (if parameter)}

\item{eps}{a named list of size 1 (see Z)}

\item{zFixed}{if TRUE zMatrix is only simulated once and stays fixed}

\item{epsFixed}{see zFixed}

\item{S}{the number of observations, only needed if zFixed or epsFixed is TRUE}
}
\value{
a function which can be used to simulate values from a factor copula model. It has the parameters theta, S and seed
}
\description{
Simulate values from a factor copula model
}
