% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFactorCopula.R
\name{fc_create}
\alias{fc_create}
\title{Simulate values from a factor copula model}
\usage{
fc_create(factor, error, beta)
}
\arguments{
\item{factor}{a configuration specified by \link[factorcopula]{config_factor}}

\item{error}{a configuation specified by \link[factorcopula]{config_error}}

\item{beta}{a parameter matrix of factor loadings specified by \link[factorcopula]{config_beta}}
}
\value{
a function which can be used to simulate values from a factor copula model. It has the parameters theta, S and seed.
}
\description{
Simulate values from a factor copula model
}
