% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm.R
\name{fc_fit}
\alias{fc_fit}
\title{Fit a factor copula model}
\usage{
fc_fit(Y, factor, error, beta, lower, upper, k = rep(1, ncol(Y)), S = 25 *
  nrow(Y), se = FALSE, B = 1000, control.first.stage = list(algorithm =
  "NLOPT_GN_DIRECT_L", stopval = 0, xtol_rel = 1e-04, maxeval = 200),
  control.second.stage = list(algorithm = "NLOPT_LN_SBPLX", stopval = 0,
  xtol_rel = 1e-16, maxeval = 10000))
}
\arguments{
\item{Y}{A dataframe or matrix like object}

\item{factor}{specification of latent variables, see \link[factorcopula]{config_factor}}

\item{error}{specification of error term, see \link[factorcopula]{config_error}}

\item{beta}{specification of parameter matrix, see \link[factorcopula]{config_beta}}

\item{lower}{Lower bound for optimazation: Named vector with parameters}

\item{upper}{Upper bound for optimazation: Named vector with parameters}

\item{k}{A vector with length ncol(Y) defining the groups (e.q. equi-dependence or block-euqidependence model)}

\item{S}{The number of simulations to use}

\item{se}{Wether to estimate standard errors and confidence intervalls}

\item{B}{number of bootstrap replication for estimating standard error}

\item{control.first.stage}{named list of options passed to the nloptr method. A full description of all options is shown by the function \code{nloptr::nloptr.print.options()}.}

\item{control.second.stage}{named list of options passed to the nloptr method. A full description of all options is shown by the function \code{nloptr::nloptr.print.options()}.}
}
\value{
a list of optimization results
}
\description{
Fit a factor copula model
}
