% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakTest.R
\name{fc_mstat}
\alias{fc_mstat}
\title{Calculate recursive test statistics for the moments based break test}
\usage{
fc_mstat(Y, tSeq, k, cl = NULL)
}
\arguments{
\item{Y}{matrix of observed values}

\item{tSeq}{A vector of positive integers}

\item{k}{a vector defining the groups of the variables}

\item{cl}{a cluster object, see \link[snow]{makeCluster}}
}
\description{
Calculate recursive test statistics for the moments based break test
}
