% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakTest.R
\name{fc_pstat}
\alias{fc_pstat}
\title{Return the p statistics of a recursive factor copula model}
\usage{
fc_pstat(theta, tSeq)
}
\arguments{
\item{theta}{A numeric matrix with length(tSeq) rows of recursive parameters}

\item{tSeq}{A vector of positive integers}
}
\value{
a vector of P statistics for each t
}
\description{
Return the p statistics of a recursive factor copula model
}
