% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm.R
\name{fitFactorCopula}
\alias{fitFactorCopula}
\title{Fit a factor copula model}
\usage{
fitFactorCopula(Y, copFun, k = rep(1, ncol(Y)), S, lower, upper,
  seed = runif(1, 1, .Machine$integer.max), method = c("DEoptim", "genoud",
  "subplex"), control)
}
\arguments{
\item{Y}{A dataframe or matrix like object}

\item{copFun}{A copula function estimated by @references factorCopula}

\item{k}{A vector with length ncol(Y) defining the groups (e.q. equi-dependence or block-euqidependence model)}

\item{S}{The number of simulations to use}

\item{lower}{Lower bound for optimazation: Named vector with parameters}

\item{upper}{Upper bound for optimazation: Named vector with parameters}

\item{seed}{Fixed random number seed to use during optimazation}

\item{method}{"DEoptim", "genoud" or "subplex" (see Details)}

\item{control}{named list of arguments passed to the optimizer}
}
\value{
a list of optimazation results
}
\description{
Fit a factor copula model
}
