% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMM.R, R/clusterUtils.R, R/utils.R
\name{loadPackagesOnCluster}
\alias{loadPackagesOnCluster}
\alias{loadPackagesOnCluster}
\alias{loadPackagesOnCluster}
\title{Load packages on a cluster}
\usage{
loadPackagesOnCluster(cl, packages)

loadPackagesOnCluster(cl, packages)

loadPackagesOnCluster(cl, packages)
}
\arguments{
\item{cl}{cluster object created by @references snow::makeCluster}

\item{packages}{list or vector of package names}

\item{cl}{cluster object created by @references snow::makeCluster}

\item{packages}{list or vector of package names}

\item{cl}{cluster object created by @references snow::makeCluster}

\item{packages}{list or vector of package names}
}
\value{
TRUE if all packages could be loaded on all cluster nodes

TRUE if all packages could be loaded on all cluster nodes

TRUE if all packages could be loaded on all cluster nodes
}
\description{
Load packages on a cluster

Load packages on a cluster

Load packages on a cluster
}
