\name{PTSD}
\alias{PTSD}
\docType{data}
\title{
The Post-traumatic stress disorder (PTSD)
}
\description{
The measure consists of 20 categorical items divided into four domains: (1) intrusions (e.g., repeated, dis- turbing, and unwanted memories), (2) avoidance (e.g., avoiding external reminder of the stressful experience), (3) cognition and mood alterations (e.g., trouble remembering important parts of the stressful experience) and (4) reactivity alterations (e.g., taking too many risks or doing things that could cause you harm). Each item is answered in a five-point ordinal scale: ``0 = Not at all'', ``1 = A little bit'', ``2 = Moderately'', ``3 = Quite a bit'', and ``4 = Extremely''. 
}
\usage{data(PTSD)}
\format{
  A data frame with 221 observations and  20 items.
}

\source{
Williams, D. and Mulder, J. (2020).
\emph{BGGM: Bayesian Gaussian Graphical Models}.
R package version 1.0.0.
}

\keyword{datasets}
