% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFuns.R
\name{rst}
\alias{rst}
\title{Generate random numbers from the skewed t distribution}
\usage{
rst(n, nu = 1e+09, lambda = 0)
}
\arguments{
\item{n}{number of observations}

\item{nu}{degree of freedoms, must be > 2}

\item{lambda}{skewness factor between (-1:1)}
}
\value{
a vactor of length n
}
\description{
Generate random numbers from the skewed t distribution
}
