% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFactorCopula.R
\name{config_beta}
\alias{config_beta}
\title{Configurate the loadings of a factorcopula model}
\usage{
config_beta(k, Z = NULL)
}
\arguments{
\item{k}{Numeric vector of length N with eventually positive increasing integers from 1 to N.}

\item{Z}{Number of latent variables}
}
\value{
A character matrix of parameters to be used in \link[factorcopula]{fc_create} or \link[factorcopula]{fc_fit}
}
\description{
Configurate the loadings of a factorcopula model
}
