% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakTest.R
\name{fc_critval}
\alias{fc_critval}
\title{Simulate critival values for either the copula or the moments based break test}
\usage{
fc_critval(type = c("moments", "copula"), Y, B, tSeq, k, copFun = NULL,
  theta = NULL, cl = NULL)
}
\description{
Simulate critival values for either the copula or the moments based break test
}
