% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm.R
\name{fc_fit}
\alias{fc_fit}
\title{Fit a factor copula model}
\usage{
fc_fit(Y, config_factor, config_error, config_beta, lower, upper, recursive,
  control, S, k, cl = NULL, trials = length(cl), load.balancing = TRUE)
}
\arguments{
\item{Y}{A dataframe or matrix like object}

\item{lower}{Lower bound for optimazation: Named vector with parameters}

\item{upper}{Upper bound for optimazation: Named vector with parameters}

\item{recursive}{Wether to estimate recursive or just full model}

\item{control}{named list of arguments passed to the Subplex Algorithm (see nl.opts for help)}

\item{S}{The number of simulations to use}

\item{k}{A vector with length ncol(Y) defining the groups (e.q. equi-dependence or block-euqidependence model)}

\item{copFun}{A copula function estimated by @references factorCopula}
}
\value{
a matrix or vector of parameters
}
\description{
Fit a factor copula model
}
