% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm.R
\name{fc_fit}
\alias{fc_fit}
\title{Fit a factor copula model}
\usage{
fc_fit(Y, copFun, lower, upper, method = c("DEoptim", "genoud", "subplex",
  "two-stage"), control, cluster = NULL, trials = NULL, S = NULL,
  k = rep(1, ncol(Y)), seed = runif(1, 1, .Machine$integer.max))
}
\arguments{
\item{Y}{A dataframe or matrix like object}

\item{copFun}{A copula function estimated by @references factorCopula}

\item{lower}{Lower bound for optimazation: Named vector with parameters}

\item{upper}{Upper bound for optimazation: Named vector with parameters}

\item{method}{"DEoptim", "genoud" or "subplex" (see Details)}

\item{control}{named list of arguments passed to the optimizer}

\item{S}{The number of simulations to use}

\item{k}{A vector with length ncol(Y) defining the groups (e.q. equi-dependence or block-euqidependence model)}

\item{seed}{Fixed random number seed to use during optimazation}
}
\value{
a list of optimazation results
}
\description{
Fit a factor copula model
}
