% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakTest.R
\name{fc_mstat}
\alias{fc_mstat}
\title{Calculate recursive test statistics for the moments based break test}
\usage{
fc_mstat(Y, tSeq, k = rep(1, ncol(Y)), cl = NULL)
}
\description{
Calculate recursive test statistics for the moments based break test
}
